+++ src/main/java/com/danit/fs12/entity/user/User.java
@@ -5,12 +5,12 @@ import com.danit.fs12.entity.bookmark.Bookmark;
 import com.danit.fs12.entity.certification.Certification;
 import com.danit.fs12.entity.chat.Chat;
 import com.danit.fs12.entity.comment.Comment;
-import com.danit.fs12.entity.commentLike.CommentLike;
+import com.danit.fs12.entity.commentlike.CommentLike;
 import com.danit.fs12.entity.education.Education;
 import com.danit.fs12.entity.group.Group;
 import com.danit.fs12.entity.message.Message;
 import com.danit.fs12.entity.post.Post;
-import com.danit.fs12.entity.postLike.PostLike;
+import com.danit.fs12.entity.postlike.PostLike;
 import com.danit.fs12.entity.workplace.WorkPlace;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.AllArgsConstructor;
