+++ src/main/java/com/danit/fs12/service/UserService.java
@@ -1,9 +1,9 @@
 package com.danit.fs12.service;
 
 import com.danit.fs12.entity.comment.Comment;
-import com.danit.fs12.entity.commentLike.CommentLike;
+import com.danit.fs12.entity.commentlike.CommentLike;
 import com.danit.fs12.entity.post.Post;
-import com.danit.fs12.entity.postLike.PostLike;
+import com.danit.fs12.entity.postlike.PostLike;
 import com.danit.fs12.entity.user.User;
 import com.danit.fs12.exception.ForbiddenException;
 import com.danit.fs12.repository.CommentRepository;
